[fix] payment/bff 패키지 -> payment 패키지로 통합 #182
