+++ payment/src/main/java/com/pyokemon/payment/controller/PaymentController.java
@@ -3,13 +3,12 @@ package com.pyokemon.payment.controller;
 import java.util.List;
 import java.util.Map;
 
+import com.pyokemon.common.dto.IdsRequest;
+import com.pyokemon.payment.dto.*;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import com.pyokemon.payment.dto.PaymentConfirmRequestDto;
-import com.pyokemon.payment.dto.PaymentConfirmResponseDto;
-import com.pyokemon.payment.dto.PaymentInitiateRequestDto;
 import com.pyokemon.payment.service.PaymentService;
 import com.pyokemon.payment.service.TossPaymentService;
 
@@ -43,6 +42,24 @@ public class PaymentController {
       return ResponseEntity.badRequest().build();
 
     }
+  }
+
+  @GetMapping("/{paymentId}")
+  public PaymentInfoDto getPayment(@PathVariable Long paymentId) {
+    return paymentService.getPayment(paymentId);
+  }
+
+  @PostMapping("/_batch")
+  public List<PaymentInfoDto> getPayments(@RequestBody IdsRequest request) {
+    return paymentService.getPayments(request.getIds());
+  }
+
+  @GetMapping("/summary/revenue")
+  public ResponseEntity<TotalRevenueResponseDto> getTotalRevenue(
+          @RequestParam("scheduleIds") List<Long> scheduleIds) {
+
+    TotalRevenueResponseDto responseDto = paymentService.getTotalRevenue(scheduleIds);
+    return ResponseEntity.ok(responseDto);
   }
 
 }
