+++ payment/src/main/java/com/pyokemon/payment/dto/TotalRevenueResponseDto.java
@@ -1,4 +1,4 @@
-package com.pyokemon.payment.bff.dto;
+package com.pyokemon.payment.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
