+++ payment/src/main/java/com/pyokemon/payment/repository/PaymentRepository.java
@@ -4,6 +4,10 @@ import org.apache.ibatis.annotations.Mapper;
 
 import com.pyokemon.payment.dto.PaymentDto;
 import com.pyokemon.payment.entity.Payment;
+import org.apache.ibatis.annotations.Param;
+
+import java.util.List;
+import java.util.Optional;
 
 @Mapper
 public interface PaymentRepository {
@@ -18,4 +22,10 @@ public interface PaymentRepository {
   void cancelPayment(String orderId, String status);
 
   Payment selectLatestByBookingId(Long bookingId);
+
+  Optional<Payment> findByPaymentId(@Param("paymentId") Long paymentId);
+
+  List<Payment> findByPaymentIds(@Param("ids") List<Long> ids);
+
+  Long sumTotalRevenueByScheduleIds(@Param("scheduleIds") List<Long> scheduleIds);
 }
