--- payment/src/main/resources/mapper/PaymentBffMapper.xml
@@ -1,45 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<!DOCTYPE mapper
-        PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
-        "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-
-<mapper namespace="com.pyokemon.payment.bff.repository.PaymentBffRepository">
-
-    <select id="findByPaymentId" resultType="Payment">
-        SELECT * FROM tb_payment WHERE payment_id = #{paymentId}
-    </select>
-
-    <!-- 배치 조회 -->
-    <select id="findByPaymentIds"
-            parameterType="map"
-            resultType="com.pyokemon.payment.entity.Payment">
-        SELECT payment_id, account_id, amount, method, status, created_at, updated_at
-        FROM tb_payment
-        WHERE
-        <choose>
-            <when test="ids != null and ids.size() > 0">
-                payment_id IN
-                <foreach collection="ids" item="id" open="(" separator="," close=")">
-                    #{id}
-                </foreach>
-            </when>
-            <otherwise>
-                1 = 0
-            </otherwise>
-        </choose>
-    </select>
-
-    <select id="sumTotalRevenueByScheduleIds" resultType="long">
-        SELECT
-        IFNULL(SUM(amount), 0)
-        FROM
-        tb_payment
-        WHERE
-        status = 'DONE'
-        AND event_schedule_id IN
-        <foreach collection="scheduleIds" item="id" open="(" separator="," close=")">
-            #{id}
-        </foreach>
-    </select>
-
-</mapper>
